+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 * support hash rocket syntax for gem and gemspec options
+* improve group parsing to handle comments and trailing tokens correctly
 
 ## [0.8.0](https://github.com/contriboss/gemfile-go/compare/v0.7.3...v0.8.0) (2026-01-22)